# [WIP, autodiff] use sparse differentiation for rotation matrix