--- a/math/diff_matrix.cc
+++ b/math/diff_matrix.cc
@@ -1,0 +1,4 @@
+// For now, this is an empty .cc file that only serves to confirm
+// differentiable_matrix.h is a stand-alone header.
+
+#include "drake/math/diff_matrix.h"