… math/diff_matrix.h
… math/test/diff_matrix_test.cc