(do not review) This will be a revised & hardened version of Russ's prototype in #17236. Should be a performance improvement only.

Adds a new DifferentiableMatrix type that holds a Matrix and its non-zero derivatives.
Uses a 3x3 instance of that matrix as the representation for RotationMatrix.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/robotlocomotion/drake/17335)
<!-- Reviewable:end -->