@sherm1 2022-07-11
WIP: Autodiff benchmark timings (rough, times in μs)

|                          | master  | Russ prototype | This PR | Speedup
| -----------------| ---------|------------------|---------| ----------
| pos_constraint | 1810      | 963                   | 562      |  3.2X
| cassie mass     | 1336      | 1037                  | 955      | 1.4X
| cassie invdyn  |  2111     | 1426                  | 1339    | 1.6X
| cassie fwddyn | 3013      | 2674                  | 2444    | 1.2X

These are just the best times I saw after repeated runs on my Ubuntu VM which has high variability. I'll take some more reliable measurements when I can. (Note that Jeremy's #17492 rewrite of AutoDiff may make this moot.)