@sherm1 2022-07-13
WIP: Autodiff benchmark timings in μs
These are from my trusty Puget, with much less variability. The speedups are effectively identical to the ones I measured on my VM.

|                          | master  | Russ prototype | This PR | Speedup
| -----------------| ---------|------------------|---------| ----------
| pos_constraint | 2596      | 1594                 | 843      |  3.1X
| cassie mass     | 2040      | 1659                  | 1397    | 1.5X
| cassie invdyn  |  2980     | 2233                  | 1787    | 1.7X
| cassie fwddyn | 4610      | 4181                  | 3779    | 1.2X

Next step is to try using AVX matrix multiplies for the 3x3 case used by RotationMatrix.