@jwnimmer-tri 2022-07-13
Note that until we get the zeros out of the benchmark (#17539), I don't believe _any_ cassie autodiff benchmarks are valid measurements of an auatodiff implementation that special-cases for zero derivatives (either this PR, or #17492).  We could easily be driving all gradients to zeros in the benchmarks (anytime we multiply by a*b, e.g. if a==0 then we zero out b's partials).